PIC: Use split irqchip emulations
Assmue kernel_irqchip=split.
Use PIC/IOAPIC userspace emulations and LAPIC kernel emulation.
