+++ hw/i386/kvm/Makefile.objs
@@ -1,1 +1,1 @@
-obj-y += clock.o apic.o i8259.o ioapic.o i8254.o
+obj-y += clock.o apic.o i8254.o
